+++ app/cashboard/(overview)/page.tsx
@@ -1,0 +1,22 @@
+import { ChartAreaInteractive } from '@/components/chart-area-interactive';
+import { DataTable } from '@/components/data-table';
+import { SectionCards } from '@/components/section-cards';
+
+import data from '@/app/cashboard/data.json';
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'My Cashboard',
+};
+
+export default function Page() {
+  return (
+    <div className='flex flex-col gap-4 py-4 md:gap-6 md:py-6'>
+      <SectionCards />
+      <div className='px-4 lg:px-6'>
+        <ChartAreaInteractive />
+      </div>
+      <DataTable data={data} />
+    </div>
+  );
+}
